requirements.txt
aiogram
python-dotenv
mysql-connector-python
aiohttp
requests  # Додано для роботи з CoinGecko